Refer to the element by className (#301)
It can be useful when we have few elements and need to scroll to first of them. Right now as I know, `react-scroll` can't handle this situation except elements with the name attribute (so it can be one of the form elements) or by handling complex logic on the client side to keep tracking of multiple ids of elements.